Test the returned type in the OER getInstance identity guards: UINT32.getInstance and etsi102941 Version.getInstance guarded on UINT8, EtsiTs103097DataEncryptedUnicast.getInstance on its sibling EtsiTs103097DataEncrypted, and OEROptional.getObject had its isInstance arguments transposed, so the first two rejected their own type and the others threw ClassCastException or resolved every optional field reflectively, relates to github #2373.

+++ util/src/main/java/org/bouncycastle/oer/its/ieee1609dot2/basetypes/UINT32.java
@@ -32,7 +32,7 @@ public class UINT32
 
     public static UINT32 getInstance(Object o)
     {
-        if (o instanceof UINT8)
+        if (o instanceof UINT32)
         {
             return (UINT32)o;
         }
